이동 정보 전송

+++ src/main/java/com/inha/endgame/room/rVector3D.java
@@ -8,11 +8,11 @@ import java.io.Serializable;
 public class rVector3D implements Serializable {
     private static final long serialVersionUID = 1151561654154545L;
 
-    private double x;
-    private double y;
-    private double z;
+    private float x;
+    private float y;
+    private float z;
 
-    public rVector3D(double x, double y, double z) {
+    public rVector3D(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
